@@ -8,24 +8,27 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "autoprefixer": "^9.8.8",
     "axios": "^0.21.1",
     "core-js": "^3.8.3",
     "jsonapi-vuex": "^4.4.2",
     "lazysizes": "^5.3.0",
-    "vue": "^2.6.11a",
-    "vue-parallaxy": "^1.1.1",
-    "vue-router": "^3.5.1",
-    "vuex": "^3.6.2"
+    "postcss": "^7.0.39",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
+    "tailwindcss-textshadow": "^2.1.3",
+    "vue": "^3.2.21",
+    "vue-router": "^4.0.12",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^4.5.11",
     "@vue/cli-plugin-eslint": "^4.5.11",
     "@vue/cli-plugin-router": "^4.5.11",
-    "@vue/cli-service": "^4.5.11",
+    "@vue/cli-service": "^4.5.13",
+    "@vue/compiler-sfc": "^3.2.11",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0-alpha.0"
   },
   "eslintConfig": {
     "root": true,
@@ -33,7 +36,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
